Acceso_a_Datos/PR8 - Acceso Biométrico/DAO Biopass/requirements.txt
# lista de herramientas externas que necesito

python-dotenv
psycopg2-binary
opencv-python
Pillow
